+++ vllm_ascend/platform.py
@@ -31,10 +31,10 @@ from vllm_ascend.config.vllm_ascend import init_ascend_config
 # isort: off
 from vllm_ascend.utils import (ASCEND_QUANTIZATION_METHOD,
                                COMPRESSED_TENSORS_METHOD, COMPILATION_PASS_KEY,
-                               AscendDeviceType, get_ascend_device_type,
-                               update_cudagraph_capture_sizes)
-from vllm_ascend.config.utils import is_vl_model, enable_sp, update_default_aclgraph_sizes, update_aclgraph_sizes, \
-    refresh_block_size, check_kv_extra_config
+                               AscendDeviceType, get_ascend_device_type)
+from vllm_ascend.config.utils import is_vl_model, enable_sp, refresh_block_size, check_kv_extra_config
+from vllm_ascend.config.aclgraph import update_cudagraph_capture_sizes, update_default_aclgraph_sizes, \
+    update_aclgraph_sizes
 
 if TYPE_CHECKING:
     from vllm.config import ModelConfig, VllmConfig
