+++ vllm_ascend/utils.py
@@ -370,86 +370,6 @@ def vllm_version_is(target_vllm_version: str):
             "format of x.y.z.")
 
 
-def get_max_hidden_layers(hf_config) -> int:
-    cfg_dict = hf_config.to_dict()
-    layer_counts = []
-
-    def _rec_find(d):
-        if isinstance(d, dict):
-            for k, v in d.items():
-                if k == "num_hidden_layers" and isinstance(v, int):
-                    layer_counts.append(v)
-                else:
-                    _rec_find(v)
-
-    _rec_find(cfg_dict)
-    if not layer_counts:
-        raise ValueError("Not found num_hidden_layers in model config.")
-    return max(layer_counts)
-
-
-# Update cudagraph capture sizes for vllm config
-def update_cudagraph_capture_sizes(vllm_config: VllmConfig,
-                                   cudagraph_capture_sizes: List[int]):
-
-    valid_max_size = (cudagraph_capture_sizes[-1]
-                      if cudagraph_capture_sizes else 0)
-    if (vllm_config.compilation_config.max_cudagraph_capture_size is not None
-            and vllm_config.compilation_config.max_cudagraph_capture_size
-            != valid_max_size):
-        if vllm_config.compilation_config.cudagraph_capture_sizes is not None:
-            raise ValueError(
-                "customized max_cudagraph_capture_size"
-                f"(={vllm_config.compilation_config.max_cudagraph_capture_size}) "
-                "should be consistent with the max value of "
-                f"cudagraph_capture_sizes(={valid_max_size})")
-        logger.warning(
-            "Truncating max_cudagraph_capture_size to %d",
-            valid_max_size,
-        )
-
-    vllm_config.compilation_config.max_cudagraph_capture_size = valid_max_size
-
-    if vllm_config.compilation_config.cudagraph_capture_sizes is not None and len(
-            cudagraph_capture_sizes) < len(
-                vllm_config.compilation_config.cudagraph_capture_sizes):
-        logger.warning(
-            ("cudagraph_capture_sizes specified in compilation_config"
-             " %s is overridden by config %s"),
-            vllm_config.compilation_config.cudagraph_capture_sizes,
-            cudagraph_capture_sizes,
-        )
-    vllm_config.compilation_config.cudagraph_capture_sizes = cudagraph_capture_sizes
-    vllm_config.compilation_config.post_init_cudagraph_sizes()
-
-
-def _is_default_capture_sizes(vllm_config: VllmConfig) -> bool:
-    """
-    Check whether it is vLLM default capture sizes.
-    """
-
-    max_cudagraph_capture_size = \
-        vllm_config.compilation_config.max_cudagraph_capture_size
-    cudagraph_capture_sizes = [
-        i for i in [1, 2, 4] if i <= max_cudagraph_capture_size
-    ]
-    if max_cudagraph_capture_size >= 8:
-        # Step size 8 for small batch sizes, up to 256(not included)
-        cudagraph_capture_sizes += list(
-            range(8, min(max_cudagraph_capture_size + 1, 256), 8))
-    if max_cudagraph_capture_size >= 256:
-        # Step size 16 for larger batch sizes
-        cudagraph_capture_sizes += list(
-            range(256, max_cudagraph_capture_size + 1, 16))
-    # in newer version, vLLM use ascending order of cudagraph_capture_sizes.
-    target_cudagraph_capture_sizes = sorted(cudagraph_capture_sizes)
-    if target_cudagraph_capture_sizes == \
-            vllm_config.compilation_config.cudagraph_capture_sizes:
-        return True
-
-    return False
-
-
 # TODO(wxy): Move to ops module
 def dispose_tensor(x: torch.Tensor):
     x.set_(torch.empty((0, ), device=x.device, dtype=x.dtype))
